fix: divide input model
