+++ packages/vlossom/src/components/vs-select/VsSelect.vue
@@ -207,12 +207,12 @@ import {
     useColorScheme,
     useStyleSet,
     getResponsiveProps,
-    getInputProps,
     useInput,
     getInputOptionProps,
     useInputOption,
     useStateClass,
 } from '@/composables';
+import { getInputProps } from '@/models';
 import { useAutocomplete, useFocusControl, useInfiniteScroll, useSelectOption, useToggleOptions } from './composables';
 import { VsComponent, type ColorScheme } from '@/declaration';
 import { VsIcon } from '@/icons';
@@ -228,7 +228,7 @@ export default defineComponent({
     name,
     components: { VsInputWrapper, VsChip, VsIcon },
     props: {
-        ...getInputProps<any, []>(),
+        ...getInputProps<any>(),
         ...getInputOptionProps(),
         ...getResponsiveProps(),
         colorScheme: { type: String as PropType<ColorScheme> },
