+++ packages/vlossom/src/composables/__tests__/input-composable.test.ts
@@ -1,30 +1,11 @@
 import { beforeEach, afterEach, describe, expect, it, vi } from 'vitest';
 import { mount } from '@vue/test-utils';
 import { defineComponent, nextTick, ref, toRefs } from 'vue';
-import { getInputProps, useInput } from '@/composables/input-composable';
+import { getInputProps } from '@/models';
+import { useInput } from '@/composables/input-composable';
 import { UIState } from '@/declaration';
 
 import type { StateMessage } from '@/declaration';
-
-describe('getInputProps', () => {
-    it('input component에 필요한 props들을 가져올 수 있다', () => {
-        // when
-        const props = getInputProps();
-
-        // then
-        expect(props).toMatchSnapshot();
-    });
-
-    it('input props 중 제외할 props를 정할 수 있다', () => {
-        // when
-        const props = getInputProps('label', 'messages');
-
-        // then
-        expect(props).toMatchSnapshot();
-        expect(props).not.toHaveProperty('label');
-        expect(props).not.toHaveProperty('messages');
-    });
-});
 
 describe('useInput composable', () => {
     const inputValue = ref('');
@@ -37,7 +18,7 @@ describe('useInput composable', () => {
     const TestInputComponent = defineComponent({
         render: () => null,
         props: {
-            ...getInputProps<string, []>(),
+            ...getInputProps<string>(),
             modelValue: { type: String, default: '' },
         },
         setup(props, ctx) {
